pyproject.toml
[tool.poetry]
name = "synthetic-trial-data"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "synthetic_trial_data"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pandas = "<3"
sas7bdat = "^2.2.3"
jupyter = "^1.0.0"
ipykernel = "^6.24.0"
pyreadstat = "^1.2.2"
tqdm = "^4.65.0"
polars = "^0.18.7"
pyarrow = ">=12.0.1"
pyyaml = "5.3.1"
typeguard = "<=5.0.0"
scikit-learn = ">=1.0.0"
synthcity = "^0.2.6"
seaborn = "^0.12.2"
jax = "^0.4.14"
jenkspy = "^0.3.3"
geomloss = "^0.2.6"
pykeops = "^2.1.2"
umap-learn = "^0.5.3"
transformers = "4.32.0"
tokenizers = "^0.13.3"
sentencepiece = "^0.1.99"
accelerate = "^0.22.0"
datasets = "^2.14.5"
mlflow = "^2.7.1"
requests = ">=2.30, <2.31"
matplotlib= ">=3.6"
hydra-core = "^1.3.2"
optuna = "^3.3.0"
plotly = "^5.17.0"
jaxlib = "^0.4.14"
botocore = "^1.31.59"
sdv = "^1.4.0"
pydantic = "<2.0"


[tool.poetry.group.macos]
optional = true

[tool.poetry.group.macos.dependencies]
tensorflow-probability = ">=0.19.0"
tensorflow-metal = "^1.0.1"
tensorflow-macos = "^2.13.0"
torch = "<2.0"

[tool.poetry.group.linux-cpu]
optional = true

[tool.poetry.group.linux-cpu.dependencies]
torch = "^1.9.0"

[tool.poetry.group.linux-gpu]
optional = true

[tool.poetry.group.linux-gpu.dependencies]
torch = "^1.9.0+cu111"

[tool.poetry.group.linux]
optional = true

[tool.poetry.group.linux.dependencies]
tensorflow = "^2.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
